extract_to_gcs/requirements.txt
Flask
google-cloud-storage
google-cloud-bigquery
requests
gunicorn
beautifulsoup4

